+++ src/doNothing.test.ts
@@ -15,21 +15,3 @@ describe("runtime", () => {
     doesNothing(["a"]);
   });
 });
-
-describe("typing", () => {
-  test("supported in any api", () => {
-    mockApi({
-      onMixOfParams: doNothing(),
-      onNoParams: doNothing(),
-      onVariadicParams: doNothing(),
-    });
-  });
-});
-
-function mockApi(_options: {
-  readonly onMixOfParams: (result: string, isOptionalBoolean?: true) => void;
-  readonly onNoParams: () => void;
-  readonly onVariadicParams: (...args: ReadonlyArray<string>) => void;
-}): void {
-  /* do nothing */
-}
